Create a base airdrop accountant module and use it in all evm chains
