+++ rotkehlchen/chain/ethereum/accountant.py
@@ -1,5 +1,6 @@
 from typing import TYPE_CHECKING
 
+from rotkehlchen.chain.ethereum.modules.airdrops.constants import ETHEREUM_AIRDROPS_LIST
 from rotkehlchen.chain.evm.accounting.aggregator import EVMAccountingAggregator
 from rotkehlchen.user_messages import MessagesAggregator
 
@@ -14,4 +15,8 @@ class EthereumAccountingAggregator(EVMAccountingAggregator):
             node_inquirer: 'EthereumInquirer',
             msg_aggregator: 'MessagesAggregator',
     ) -> None:
-        super().__init__(node_inquirer=node_inquirer, msg_aggregator=msg_aggregator)
+        super().__init__(
+            node_inquirer=node_inquirer,
+            msg_aggregator=msg_aggregator,
+            airdrops_list=ETHEREUM_AIRDROPS_LIST,
+        )
