+++ rotkehlchen/chain/optimism/accountant.py
@@ -3,6 +3,8 @@ from typing import TYPE_CHECKING
 from rotkehlchen.chain.evm.accounting.aggregator import EVMAccountingAggregator
 from rotkehlchen.user_messages import MessagesAggregator
 
+from .constants import CPT_OPTIMISM
+
 if TYPE_CHECKING:
     from .node_inquirer import OptimismInquirer
 
@@ -14,4 +16,8 @@ class OptimismAccountingAggregator(EVMAccountingAggregator):
             node_inquirer: 'OptimismInquirer',
             msg_aggregator: 'MessagesAggregator',
     ) -> None:
-        super().__init__(node_inquirer=node_inquirer, msg_aggregator=msg_aggregator)
+        super().__init__(
+            node_inquirer=node_inquirer,
+            msg_aggregator=msg_aggregator,
+            airdrops_list=[CPT_OPTIMISM],
+        )
