--- rotkehlchen/chain/optimism/modules/airdrops/constants.py
@@ -1,1 +1,0 @@
-CPT_OPTIMISM = 'optimism'
